+++ src/chainparams.cpp
@@ -26,7 +26,7 @@ int64_t CChainParams::GetProofOfWorkReward(int nHeight, int64_t nFees) const
     int64_t nSubsidy = 0;
 
 	if(nHeight == 1)
-		nSubsidy = (NetworkID() == CChainParams::TESTNET ? 100000 : 20000000) * COIN;  // 20Mill Pre-mine on MainNet
+        nSubsidy = (NetworkID() == CChainParams::TESTNET ? 2000000 : 20000000) * COIN;  // 20Mill Pre-mine on MainNet, 2Mill pre-mine on TestNet
 
     else if(nHeight <= nLastPOWBlock)
         nSubsidy = 0;
@@ -65,8 +65,8 @@ int64_t CChainParams::GetProofOfAnonStakeReward(const CBlockIndex* pindexPrev, i
 
     // To avoid ATXOs lower than base fee, roundup reward 1 digit above base fee (2.12345678 becomes 2.124)
     // Note: anon staking rewards are only possible from V3 on and then MIN_TX_ANON is effective
-    nSubsidy -= nSubsidy % (MIN_TX_FEE * 10);
-    nSubsidy += MIN_TX_FEE * 10;
+    nSubsidy -= nSubsidy % (nMinTxFee * 10);
+    nSubsidy += nMinTxFee * 10;
 
     return nSubsidy;
 }
@@ -184,7 +184,7 @@ public:
         convertSeeds(vFixedSeeds, pnSeed, ARRAYLEN(pnSeed), nDefaultPort);
 
         nForkV2Time = 1534888800; // MAINNET V2 chain fork (GMT: Tuesday, 21. August 2018 22.00)
-        nForkV3Time = 1600000000; // MAINNET V2 chain fork (09/13/2020 @ 12:26pm (UTC))
+        nForkV3Time = 1553810400; // MAINNET V3 chain fork (GMT: Thursday, 28. March 2019 22:00)
 
         devContributionAddress = "SdrdWNtjD7V6BSt3EyQZKCnZDkeE28cZhr";
     }
@@ -217,22 +217,22 @@ public:
         nRPCPort = 36757;
         nBIP44ID = 0x80000001;
 
-        nLastPOWBlock = 110;
-        nFirstPosv2Block = 110;
+        nLastPOWBlock = 20;
+        nFirstPosv2Block = 20;
         nFirstPosv3Block = 500;
 
         bnProofOfWorkLimit = CBigNum(~uint256(0) >> 1);
         bnProofOfStakeLimit = CBigNum(~uint256(0) >> 20);
-        bnProofOfStakeLimitV2 = CBigNum(~uint256(0) >> 16);
+        bnProofOfStakeLimitV2 = CBigNum(~uint256(0) >> 42);
 
-        nStakeMinConfirmationsLegacy = 10;
-        nStakeMinConfirmations = 10;
+        nStakeMinConfirmationsLegacy = 28;
+        nStakeMinConfirmations = 45;
 
         genesis.nBits  = bnProofOfWorkLimit.GetCompact();
-        genesis.nNonce = 1001;
+        genesis.nNonce = 20;
 
         hashGenesisBlock = genesis.GetHash();
-        assert(hashGenesisBlock == uint256("0x0eaef840827189830c177c345f53a26ad87e0770b200a83d7ff6a928d725d882"));
+        assert(hashGenesisBlock == uint256("0x0a3e03a153b1713ebc1f03fefa5d013bba4d2677ae189fcb727396b98043d95c"));
 
         base58Prefixes[PUBKEY_ADDRESS]      = list_of(127).convert_to_container<std::vector<unsigned char> >();
         base58Prefixes[SCRIPT_ADDRESS]      = list_of(196).convert_to_container<std::vector<unsigned char> >();
@@ -251,7 +251,7 @@ public:
         nForkV3Time = 1546470000; // TESTNET V3 chain fork (01/02/2019 @ 11:00pm (UTC))
 
 
-        devContributionAddress = "tQuY2feSvtYogfWPbXLgqgDT2JfdZYUf7h";
+        devContributionAddress = "tSJoPZoXumJyDmGKYo9Y7SZkJvymESFYkD";
     }
     virtual Network NetworkID() const { return CChainParams::TESTNET; }
 };